State and eigen state functions
The Ce and Se functions were modified to accept the list ofthe fourier coeffiicients, the corresponding vector will be indicated in the eigen_state function  instead of doing it inside of those functions. It is d one  according to the energy level.

Two versions of a couple of functions were added while I found what would be the best way to keep moving on regarding the spatial variable "x".